Move content uri request body to a new file
